% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_lm.R
\name{denormalize_lm}
\alias{denormalize_lm}
\title{De-normalization of data}
\usage{
denormalize_lm(x, y, parameters)
}
\arguments{
\item{x}{Vector of x values (NA values not permitted).}

\item{y}{Vector of y values (NA values not permitted). Must have same length as x.}

\item{parameters}{List of parameters to use for de-normalizaton. Normally, one will use the
output of \code{normalize_lm}, which contains the list element \code{parameters}.}
}
\description{
De-normalizes data that has
}
\examples{

# Example data
n <- 50
x2 <- seq(-30,18, length = n)
x <- x2 + 31
y <- 50 + (-x2^3 -12*x2^2 + 800*x2)/300 + rnorm(n, sd = 5)

# Normalization
norm <- normalize_lm(x, y)
plot(norm$x, norm$y, main = "Normalized data")

# Add rectangle
x_new <- c(-1.5, 1.5, 1.5, -1.5)
y_new <- c(-2, -1, 2, 2)
polygon(x_new, y_new, color = "red")

# De-normalize the data
denorm <- denormalize_lm(norm$x, norm$y, norm$parameters)
plot(denorm$x, denorm$y, main = "Normalized and denormalized data")

# De-normalize the rectangle data, based on the normalization of the data
denorm_new <- denormalize_lm(x_new, y_new, norm$parameters)
polygon(denorm_new$x, denorm_new$y, border = "red")
}
