% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftcensored_ggplot.R
\name{lc_ggplot}
\alias{lc_ggplot}
\title{Plot left-justified data}
\usage{
lc_ggplot(data, size = 2, width = 0)
}
\arguments{
\item{data}{Data frame with columns named 'x', 'y_cens', 'y_aboveLOQ', and (if you use version = 2) 'y_LOQ'}

\item{size}{Width of jittering of points (to reduce overlap between points with same y value).
The default is 0, i.e. no jittering.}
}
\description{
Plot left-justified data
}
