% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftcensored_plot.R
\name{lc_plot}
\alias{lc_plot}
\title{Plot left-justified data}
\usage{
lc_plot(
  data,
  x = "x",
  y = "y",
  uncensored = "uncensored",
  threshold = "threshold",
  results = NULL,
  y_true = NULL,
  facet = NULL
)
}
\arguments{
\item{data}{Data frame with plot data columns named 'x', 'y', 'y_aboveLOQ', and (if you use version = 2) 'y_LOQ'}

\item{x}{Name of variable with 'x' data (given as a quoted string); default is 'x'}

\item{y}{Name of variable with 'y' data, i.e. uncensored values (given as a quoted string); default is 'y'}

\item{uncensored}{Name of variable with data indicating which values are uncensored (1) or censored (0); default is 'uncensored'}

\item{threshold}{Name of variable with threshold data for both uncensored and censored data (given as a quoted string); default is 'threshold'}

\item{results}{(Optional) Named list of results from lc_linear and/or lc_fixedsplines}

\item{y_true}{(Optional) Name of variable with true 'y' data for the model generating the data (if the data has been generated by simulation)}

\item{facet}{(Optional) If a list of results is provided, setting this to either "rows", "cols" or "wrap" will show a facetted plot (i.e. one subplot per plot).
The default value is NULL, i.e. no facets are used (i.e., different models are shown using different colours).}
}
\description{
Plot left-justified data
}
\examples{

# 1. Plot data only
# Observations from a chemical in cod liver, at one location  
data_prepared <- lc_prepare(subset(polybrom, station \%in\% "23B"), 
                            x = "year",
                            y = "concentration", 
                            censored = "LOQ_flag",
                            log = TRUE)
lc_plot(data_prepared)

# 2. Plot simulated data  

# Simulate data
set.seed(991)
X <- seq(from=-1, to=1, by=.025) # generating inputs
B <- t(splines::bs(X, knots=seq(-1,1,1), degree=3, intercept = TRUE)) # creating the B-splines
num_data <- length(X); num_basis <- nrow(B)
a0 <- 0.2 # intercept
a <- rnorm(num_basis, 0, 1) # coefficients of B-splines
n_param <- length(a)
Y_model <- as.vector(a0*X + a\%*\%B)   # generate expected/mean value  
Y <- Y_model + rnorm(length(X),0,.1) # adding noise
data_sim <- data.frame(x = X, y_uncensored = Y, y_true = Y_model)
# Add censoring 
data_sim$y <- data_sim$y_uncensored
data_sim$uncensored <- 1
threshold_fixed <- -0.3
sel <- data_sim$y_uncensored < threshold_fixed
data_sim$y[sel] <- NA  
data_sim$uncensored[sel] <- 0  
data_sim$threshold <- threshold_fixed
# Plotting
lc_plot(data_sim, y_true = "Y_model")

# 3. Plot with models
\dontrun{
result_linear <- lc_linear_qi(dat_sim)
result_nonlinear <- lc_fixedsplines_qi(dat_sim, knots = 3)
lc_plot(dat_sim, 
            y_true = "y_true", 
results = list(Linear = result_linear,
               Nonlinear = result_nonlinear))
}

}
