% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftcensored_prepare.R
\name{lc_prepare}
\alias{lc_prepare}
\title{Prepare data for lc_linear}
\usage{
lc_prepare(
  data,
  x = "x",
  y = "y",
  censored = "LOQ_flag",
  censored_value = "<",
  log = FALSE,
  const = 0
)
}
\arguments{
\item{data}{Data set (a data frame)}

\item{x}{Variable name for the predictor variable}

\item{y}{Variable name of the response variable (for chemical data, concentration data)}

\item{censored}{Variable name of the variable with the LOQ flag}

\item{censored_value}{Value of LOQ flag variable for values under LOQ, by default "<"}

\item{log}{Value (TRUE/FALSE) for whether to log-transform concentrations or not (TRUE by default)}

\item{const}{Constant to add before log-transformation (to avoid problems with log of zero)}
}
\description{
This function is meant for analysis of chemical concentration data as function of year. These data are given as
two columns in the data: Concentration and LOQ flag. In our case, we assume that an empty LOQ flag
indicates that the concentration is above the LOQ. A non-empty LOQ flag indicates that the concentration is below the LOQ,
and the value given in the Concentration column is actually the LOQ. This procedure organises the data so the fit for lc_linear,
and applies the standard column names needed. In addition, the LOQ is not given when the data is above LOQ.
This procedure also sets a 'mock' LOQ for uncensored data (lowest value - 1; this does not affect the analysis).
}
\examples{

# The 'polybrom' data set contains the variables 'year', 'concentration' (concentration of a 
# polybrominated substance), and 'LOQ_flag' ('<' for measurements under LOQ, the Limit Of Quantification).
# The measurements are from three different stations. See help("polybrom").

# Show the data in the original dataset:
library(ggplot2)
ggplot(polybrom, aes(year, concentration, color = LOQ_flag)) + 
  geom_point() + 
  facet_wrap(vars(station)) 

# Make the data ready for analysis
# We also choose to log-transform the data in this case 
polybrom_for_analysis <- lc_prepare(polybrom, 
                                    x = "year", 
                                    y = "concentration", 
                                    censored = "LOQ_flag",
                                    log = TRUE)
                                    
# x, y and LOQ variables are given standard names (other variables are left unchanged). 
# 'y' is set to NA for censored data:  
ggplot(polybrom_for_analysis, aes(x, y, color = factor(uncensored))) + geom_point() + facet_wrap(vars(station))

# 'threshold' is set to a random value (lower than observed values) for all uncensored data:
ggplot(polybrom_for_analysis, aes(x, threshold, color = factor(uncensored))) + geom_point() + facet_wrap(vars(station))

}
