% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftcensored_ggplot.R
\name{leftcensored_ggplot}
\alias{leftcensored_ggplot}
\title{Plot left-justified data}
\usage{
leftcensored_ggplot(data, type = 1)
}
\arguments{
\item{data}{Data frame with columns named 'x', 'y_cens', 'y_aboveLOQ', and (if you use version = 2) 'y_LOQ'}

\item{type}{Set to either 1 (don't add y_LOQ points) or 2 (add y_LOQ points). Default is 1.}
}
\description{
Plot left-justified data
}
\examples{
# Plot data
data(concentrations)
data_prepared <- leftcensored_prepare(concentrations)
gg <- leftcensored_ggplot(data_prepared, type = 1)
gg

# Add regression line to plot
result <- lm_linear(data_prepared)
plotdata <- regression_ci(data_prepared, result$model)
gg +
  geom_ribbon(data = plotdata, aes(x = x, ymin = y_lo, ymax = y_hi),
  fill = "red", alpha = 0.25) +
  geom_path(data = plotdata, aes(x = x, y = y), color = "red3", size = 1)

}
