% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftcensored_simulate.R
\name{leftcensored_simulate}
\alias{leftcensored_simulate}
\title{Simulate censored data}
\usage{
leftcensored_simulate(
  n = 100,
  intercept = 30,
  slope = -3,
  sigma = 4,
  LOQ_1 = 20,
  LOQ_2 = 5,
  LOQ_change = 4,
  plot = TRUE,
  seed = NULL
)
}
\arguments{
\item{n}{Number of simulated observations}

\item{intercept}{Intercept of the linear relationship between y and x}

\item{slope}{Slope of the linear relationship between y and x}

\item{sigma}{Standard deviance around the linear relationship (the s.d. of y for a given x)}

\item{LOQ_1}{LOQ when x <= LOQ_change}

\item{LOQ_2}{LOQ when x > LOQ_change}

\item{LOQ_change}{The level of x when LOQ changes from LOQ_1 to LOQ_2}

\item{plot}{If TRUE, makes a plot of the data}

\item{seed}{If you want the simulation to be reproducible, set seed to some integer}
}
\value{
A data frame with 5 columns is returned invisibly. The variables are 1) x; 2) y (the real, unovserved value);
3) y_cens (the observed y values, which equals y for values above LOQ, and is NA for values under LOQ);
4) y_aboveLOQ (equals 0 for values below LOQ and 1 for values above LOQ); and 5) y_LOQ (the LOQ
value for each observation).
}
\description{
Simulates data where the real value y depends linearly on x, with random variation added.
The values are censored, so we only observe the actual value above some limit LOQ.
LOQ has two different levels depending on x. This is a typical situation if y is chemical
measurements and x is time in years, as chemical methods tends to improve (decrease LOQ) over the years.
}
\examples{

# Default parameters
leftcensored_simulate()

# Modified parameters
sim <- leftcensored_simulate(slope = -1.5, LOQ_1 = 27, LOQ_2 = 20)

# The data object of the output can be directly used in lm_leftcensored 
result <- leftcensored_lm(sim$data)

}
\seealso{
\code{\link{lm_leftcensored}}
}
