% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_lm.R
\name{normalize_lm}
\alias{normalize_lm}
\title{Normalization and removal of linear trend}
\usage{
normalize_lm(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
The function returns a list with three elements: \code{x} and \code{y}, the normalized x and y variables, and
\code{parameters}, which contains the parameters needed to denormalize the data (e.g. restore the original values of x and y).
}
\description{
This function takes two variables x and y as an input, and removes the linear trend between them as well as
normalizes the resulting variables (to mean = 0 and standard deviation = 1). Removal of the linear tremd is done using lm().
}
\examples{

# Example data
n <- 50
x2 <- seq(-30,18, length = n)
x <- x + 31
y <- 50 + (-x2^3 -12*x2^2 + 800*x2)/300 + rnorm(n, sd = 5)

# Normalization
norm <- normalize_lm(x, y)
plot(norm$x, norm$y, main = "Normalized data")

# De-normalize the data
denorm <- denormalize_lm(norm$x, norm$y, norm$parameters)
plot(denorm$x, denorm$y, main = "Normalized and denormalized data")

# Compare with the original data  
plot(x, y, main = "Original data")

}
\keyword{Statistics,}
\keyword{regression}
