% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_ci.R
\name{regression_ci}
\alias{regression_ci}
\title{Get data for plotting confidence interval of the regression}
\usage{
regression_ci(data, mcmc, n = 20)
}
\arguments{
\item{data}{data frame containing x values in a column named 'x'}

\item{mcmc}{MCMC object}

\item{n}{Number of x values used (default is 20). Increase for more smoothness of the CI}
}
\value{
Matrix of n rows and 5 columns: x, y, y_lo and y_hi
}
\description{
Get data for plotting confidence interval of the regression
}
\examples{
sim <- leftcensored_simulate()
result <- lm_linear(sim$data)
plotdata <- regression_ci(sim$data, result$model)
plot(y ~ x, type = "l", data = plotdata, ylim = range(plotdata[,3:4]))
lines(y_lo ~ x, data = plotdata, lty = 2)
lines(y_hi ~ x, data = plotdata, lty = 2)
 
}
