% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_quantiles.R
\name{regression_quantiles}
\alias{regression_quantiles}
\title{Get regression quantiles for y for a number of x values within the range of x of the given data set}
\usage{
regression_quantiles(data, mcmc, n = 20)
}
\arguments{
\item{data}{data frame containing x values in a column named 'x'}

\item{mcmc}{MCMC object}

\item{n}{Number of x values used (default is 20)}
}
\value{
Matrix of n rows and 5 columns, containing quantiles for a given x value
}
\description{
Get regression quantiles for y for a number of x values within the range of x of the given data set
}
\examples{
sim <- lc_simulate()
result <- lc_linear(sim$data)
regression_quantiles(sim$data, result$model)

}
